# docs/swagger.yaml
basePath: /api
definitions:
  app.Response:
    properties:
      code:
        example: 200
        type: integer
      data:
        type: object
      msg:
        example: success
        type: string
    type: object
host: ms.csie.org
info:
  contact:
    email: jameshwc@gmail.com
    name: jameshwc
    url: https://jameshsu.csie.org
  description: To add a level into database
  license:
    name: GPL v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  title: Add a level API
  version: "1.0"
paths:
  /game/levels/{id}:
    post:
      consumes:
      - text/plain
      description: Get a level
      parameters:
      - description: id of the level
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Get a level
      tags:
      - game
  /game/levels/new:
    post:
      consumes:
      - multipart/form-data
      description: Add a new level
      parameters:
      - description: id of the song, should have many
        in: formData
        name: songs
        required: true
        type: string
      - description: title of the level
        in: formData
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Add a new level
      tags:
      - game
  /game/lyrics/{song_id}:
    post:
      consumes:
      - text/plain
      description: Get lyrics with a song's ID; normally it is only for internal use.
      parameters:
      - description: id of the song
        in: path
        name: song_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Get lyrics with a song's ID
      tags:
      - game
  /game/songs/{id}:
    get:
      consumes:
      - text/plain
      description: Get a song instance, with a miss lyrics id randomly generated
      parameters:
      - description: id of the song
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Get a song Instance
      tags:
      - game
  /game/songs/new:
    post:
      consumes:
      - multipart/form-data
      description: Add a song to database
      parameters:
      - description: subtitle file
        in: formData
        name: file
        required: true
        type: file
      - description: youtube url
        in: formData
        name: url
        required: true
        type: string
      - description: name of the song
        in: formData
        name: name
        required: true
        type: string
      - description: singer of the song
        in: formData
        name: singer
        required: true
        type: string
      - description: miss lyrics index, seperated by commas
        in: formData
        name: miss_lyrics
        required: true
        type: string
      - description: genre of the song
        in: formData
        name: genre
        type: string
      - description: language of the song, needs to be short, and seperated by commas if multiple languages
        in: formData
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Add a song
      tags:
      - game
  /game/tours/{id}:
    get:
      consumes:
      - text/plain
      description: Add a new level
      parameters:
      - description: id of the game
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Add a new level
      tags:
      - game
schemes:
- http
swagger: "2.0"