# routers/api/game/tour.go
package game

import (
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
	"github.com/jameshwc/Million-Singer/model"
	"github.com/jameshwc/Million-Singer/pkg/app"
	"github.com/jameshwc/Million-Singer/pkg/constant"
)

// GetTour godoc
// @Summary Add a new level
// @Description Add a new level
// @Tags game
// @Accept plain
// @Produce json
// @Param id path int true "id of the game"
// @Success 200 {object} app.Response
// @Failure 400 {object} app.Response
// @Failure 404 {object} app.Response
// @Failure 500 {object} app.Response
// @Router /game/tours/{id} [get]
func GetTour(c *gin.Context) {
	appG := app.Gin{C: c}
	tourID, err := strconv.Atoi(c.Param("id"))
	if err != nil {
		appG.Response(http.StatusBadRequest, constant.INVALID_PARAMS, nil)
		return
	}
	if g, err := model.GetTour(tourID); err != nil {
		appG.Response(http.StatusInternalServerError, constant.ERROR_GET_TOUR_FAIL, nil)
	} else {
		appG.Response(http.StatusOK, constant.SUCCESS, g)
	}
}